Refactor import paths for GeneralError to use consistent casing and create global.types.ts file
