+++ src/shared/types/global.types.ts
