+++ src/shared/utils/handleErrorForm.ts
@@ -1,4 +1,4 @@
-import { YupError } from "../types/Global"
+import { YupError } from "../types/global.types"
 
 export const handleErrorForm = (error: unknown): YupError => {
   const newError = error as YupError
